Fixed Float - Int clash

+++ src/main/java/org/sonarsource/plugins/rtechnicaldebt/measures/cumulative/ComputeAverageAMC.java
@@ -27,14 +27,14 @@ public class ComputeAverageAMC implements MeasureComputer {
     // measure is already defined on files by {@link SetSizeOnFilesSensor}
     // in scanner stack
     if (context.getComponent().getType() != Component.Type.FILE) {
-      float sum = 0;
+      double sum = 0;
       int count = 0;
       for (Measure child : context.getChildrenMeasures(AVERAGE_METHOD_COMPLEXITY.key())) {
-        sum += child.getIntValue();
+        sum += child.getDoubleValue();
         count++;
       }
-      float average = count == 0 ? 0 : sum / count;
-      context.addMeasure(AVERAGE_METHOD_COMPLEXITY.key(), average);
+      double average = count == 0 ? 0 : sum / count;
+      context.addMeasure(AVERAGE_METHOD_COMPLEXITY.key(), (float) average);
     }
   }
 }
